rename director classes
